+++ sdk-ts/src/index.ts
@@ -11,4 +11,5 @@ export * from "./encoding.js";
 export * from "./identity.js";
 export * from "./client.js";
 export * from "./ids.js";
+export * from "./subscription.js";
 export * as builders from "./builders.js";